refactor: last batch of ts conversion
